+++ packages/perseus/src/widgets/__testdata__/dropdown.testdata.ts
@@ -1,4 +1,4 @@
-import { arrayOfLength, randomBoolean, randomInteger, randomSentence } from "../../../../perseus-core/src/__stories__/randomizers";
+import { arrayOfLength, randomBoolean, randomInteger, randomSentence } from "./randomizers";
 import type {PerseusRenderer} from "../../perseus-types";
 
 export const question1: PerseusRenderer = {
