+++ packages/perseus/src/widgets/__testdata__/explanation.testdata.ts
@@ -1,4 +1,4 @@
-import { randomBoolean, randomSentence } from "../../../../perseus-core/src/__stories__/randomizers";
+import { randomBoolean, randomSentence } from "./randomizers";
 import type {PerseusRenderer} from "../../perseus-types";
 
 export const question1: PerseusRenderer = {
