+++ packages/perseus/src/widgets/__testdata__/expression.testdata.ts
@@ -7,7 +7,7 @@ import {
     randomLetter,
     randomSentence,
     randomWord,
-} from "../../../../perseus-core/src/__stories__/randomizers";
+} from "./randomizers";
 import type {
     PerseusExpressionWidgetOptions,
     Version,
